+++ src/main/java/org/netmelody/cieye/server/configuration/avatar/SimpleHttp.java
@@ -1,0 +1,16 @@
+package org.netmelody.cieye.server.configuration.avatar;
+
+
+public final class SimpleHttp implements PictureUrlProvider {
+
+    @Override
+    public String imageUrlFor(String picture) {
+        return "http:" + picture;
+    }
+
+    @Override
+    public String handlesPrefix() {
+        return "http";
+    }
+
+}