+++ src/test/java/org/netmelody/cieye/server/configuration/test/RecordedKnownOffendersTest.java
@@ -14,9 +14,9 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.TemporaryFolder;
 import org.netmelody.cieye.core.domain.Sponsor;
-import org.netmelody.cieye.server.configuration.Gravatar;
 import org.netmelody.cieye.server.configuration.RecordedKnownOffenders;
 import org.netmelody.cieye.server.configuration.SettingsFile;
+import org.netmelody.cieye.server.configuration.avatar.Gravatar;
 
 
 public final class RecordedKnownOffendersTest {
@@ -80,7 +80,7 @@ public final class RecordedKnownOffendersTest {
     
     @Test public void
     looksUpOffenderAndProvidesLinkToGravatarMugshot() {
-        String gravatarUrlWithHashedEmail = new Gravatar().getUrl("john.doe@gmail.com");
+        final String gravatarUrlWithHashedEmail = new Gravatar().imageUrlFor("john.doe@gmail.com");
         assertThat(offenders.search("john"), contains(new Sponsor("", gravatarUrlWithHashedEmail)));
     }
 }