+++ src/test/java/org/netmelody/cieye/server/configuration/test/RecordedObservationTargetsTest.java
@@ -41,7 +41,8 @@ public final class RecordedObservationTargetsTest {
         assertThat(features, contains(new Feature("Jenkins 1", "http://jenkinsurl", new CiServerType("JENKINS")),
                                       new Feature("Hudson 1", "http://hudsonurl", new CiServerType("HUDSON")),
                                       new Feature("TeamCity 1", "http://teamcityurl", new CiServerType("TEAMCITY")),
-                                      new Feature("", "http://allurl", new CiServerType("JENKINS"))));
+                                      new Feature("", "http://allurl", new CiServerType("JENKINS")),
+                                      new Feature("TeamCity 2", "http://tcsecure", new CiServerType("TEAMCITY"), "user", "pass")));
     }
 
     @Test public void
@@ -56,7 +57,7 @@ public final class RecordedObservationTargetsTest {
         final Collection<Feature> features = targets.landscapes().landscapeNamed("Public Live").features();
         assertThat(features, contains(new Feature("Jenkins core", "http://ci.jenkins-ci.org", new CiServerType("JENKINS")),
                                       new Feature("Main (trunk, branches, and alternative builds)", "http://hudson.magnolia-cms.com", new CiServerType("HUDSON")),
-                                      new Feature("Gradle", "http://teamcity.jetbrains.com", new CiServerType("TEAMCITY")),
+                                      new Feature("Apache Ivy", "http://teamcity.jetbrains.com", new CiServerType("TEAMCITY")),
                                       new Feature("CI-Eye", "http://teamcity.codebetter.com", new CiServerType("TEAMCITY"))));
     }
 }